# std: Make the DEFLATE decompression routine 3x faster